bug fix_4

+++ src/main/java/me/kirito5572/commands/music/SearchCommand.java
@@ -72,34 +72,43 @@ public class SearchCommand implements ICommand {
                     }
                     try {
                         event.getChannel().retrieveMessageById(event.getChannel().getLatestMessageId()).queue(lambdaMessage1 -> {
-                            int a = Integer.parseInt(lambdaMessage1.getContentRaw());
-                            if (!audioManager.isConnected()) {
-                                audioManager.openAudioConnection(voiceChannel);
-                                Thread thread = new Thread(() -> {
-                                    AudioManager audioManager1 = event.getGuild().getAudioManager();
-                                    PlayerManager playerManager = PlayerManager.getInstance();
-                                    GuildMusicManager musicManager = playerManager.getGuildMusicManager(event.getGuild());
-                                    while (true) {
-                                        try {
-                                            Thread.sleep(1000);
-                                        } catch (InterruptedException e) {
-                                            e.printStackTrace();
-                                        }
-                                        if (!audioManager1.isConnected()) {
-                                            break;
-                                        }
-                                        if (!musicManager.player.isPaused()) {
-                                            assert voiceChannel != null;
-                                            JoinCommand.autoPaused(event, audioManager, voiceChannel, musicManager);
-                                        }
-                                    }
-                                });
-                                thread.start();
+                            int a = 0;
+                            boolean pass;
+                            try {
+                                a = Integer.parseInt(lambdaMessage1.getContentRaw());
+                                pass = false;
+                            } catch (NumberFormatException e) {
+                                pass = true;
                             }
-                            PlayerManager manager = PlayerManager.getInstance();
-                            message.delete().queue();
-                            channel.sendMessage("노래가 추가되었습니다.").queue(message1 -> message1.delete().queueAfter(5, TimeUnit.SECONDS));
-                            manager.loadAndPlay(channel, "https://youtu.be/" + data[a - 1][1]);
+                            if(!pass) {
+                                if (!audioManager.isConnected()) {
+                                    audioManager.openAudioConnection(voiceChannel);
+                                    Thread thread = new Thread(() -> {
+                                        AudioManager audioManager1 = event.getGuild().getAudioManager();
+                                        PlayerManager playerManager = PlayerManager.getInstance();
+                                        GuildMusicManager musicManager = playerManager.getGuildMusicManager(event.getGuild());
+                                        while (true) {
+                                            try {
+                                                Thread.sleep(1000);
+                                            } catch (InterruptedException e) {
+                                                e.printStackTrace();
+                                            }
+                                            if (!audioManager1.isConnected()) {
+                                                break;
+                                            }
+                                            if (!musicManager.player.isPaused()) {
+                                                assert voiceChannel != null;
+                                                JoinCommand.autoPaused(event, audioManager, voiceChannel, musicManager);
+                                            }
+                                        }
+                                    });
+                                    thread.start();
+                                }
+                                PlayerManager manager = PlayerManager.getInstance();
+                                message.delete().queue();
+                                channel.sendMessage("노래가 추가되었습니다.").queue(message1 -> message1.delete().queueAfter(5, TimeUnit.SECONDS));
+                                manager.loadAndPlay(channel, "https://youtu.be/" + data[a - 1][1]);
+                            }
                         });
                         return;
